testing: promote the cache-poisoning demo to a conformance test
The `make jailbreak` demo proved the escape by eyeballing stdout, and
lived outside the suite that CI actually runs. Reshape it into a
self-checking conformance test under testing/conformance/isolation/ so
the regression is guarded on every run.

The test stages the same escape -- translate a victim block, patch its
cached NOP into a syscall via /proc/self/maps, re-run it -- but routes
the smuggled write(2) to a pipe the parent inspects, and runs the attempt
in a child so the store fault can't take the harness down. It passes both
natively (no cache mapping to patch) and under a fixed runtime (the store
faults), and fails against the vulnerable one.
